Allow composer to be managed by other tools
We handle composer in other roles and its installed / updated via
a user with higher privileges than the user the sites managed by. We
don't want the user the project is deployed as to be able to deal with
it. This is a backwards compatible change and seems to work.