Fix order of tests for managing composer
When symfony_project_manage_composer is False, there is an error  'dict object' has no attribute 'stat'. Because when it is false you don't check if it exists, so composer_file has no .stat. But it is tested first so fails with an error.

Swapping the order of the tests fixes the issue.